api/requirements.txt
Flask
flask-cors
transformers==4.40.0
torch
Pillow
